+++ packages/atlas/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.0.20 (2024-01-23)
 
+- Renamed `satellite` parameter to `lightSatellite` in `MapStyle`
 - Added `darkSatellite` parameter to `MapStyle`
 
 ## v1.0.19 (2024-01-16)